- Исправления 1.17.1 - Начинаем спринт 17. Продолжаем причесывать проект по mvvm.
