--- Tracker/Protocols/CloseScreenDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  File.swift
-//  Tracker
-//
-//  Created by Kirill Sklyarov on 02.04.2024.
-//
-
-import Foundation
-
-protocol CloseScreenDelegate: AnyObject {
-    func closeFewVCAfterCreatingTracker()
-}
