+++ Tracker/ViewControllers/CreatingNewTrackerVC/ViewModel/CreatingNewTrackerViewModel.swift
@@ -7,62 +7,84 @@
 
 import Foundation
 
-final class CreatingNewTrackerViewModel {
+final class CreatingNewTrackerViewModel: CreatingNewTrackerViewModelProtocol {
+   
+    var tableViewRows = ["Категория", "Расписание"]
     
-    let rowHeight = CGFloat(75)
-    
-    var tableViewHeight: CGFloat {
-        rowHeight * CGFloat(tableViewRows.count)
-    }
-    
-    
-    let tableViewRows = ["Категория", "Расписание"]
-    
-    let arrayOfEmoji = ["🙂","😻","🌺","🐶","❤️","😱",
+    var arrayOfEmoji = ["🙂","😻","🌺","🐶","❤️","😱",
                         "😇","😡","🥶","🤔","🙌","🍔",
                         "🥦","🏓","🥇","🎸","🏝️","😪",]
     
-    let arrayOfColors = ["#FD4C49", "#FF881E", "#007BFA", "#6E44FE", "#33CF69", "#E66DD4",
+    var arrayOfColors = ["#FD4C49", "#FF881E", "#007BFA", "#6E44FE", "#33CF69", "#E66DD4",
                          "#F9D4D4", "#34A7FE", "#46E69D", "#35347C", "#FF674D", "#FF99CC",
                          "#F6C48B", "#7994F5", "#832CF1", "#AD56DA", "#8D72E6", "#2FD058"]
     
     let coreDataManager = TrackerCoreManager.shared
+        
+    var trackerName: String? {
+        didSet {
+            isDoneButtonEnable?()
+        }
+    }
     
-    var newTaskName: String?
-    var selectedEmoji: String?
-    var selectedColor: String?
-    var selectedCategory: String?
-    var selectedSchedule: String?
+    var selectedEmoji: String? {
+        didSet {
+            isDoneButtonEnable?()
+        }
+    }
     
-    var informAnotherVCofCreatingTracker: ( () -> Void )?
+    var selectedColor: String? {
+        didSet {
+            isDoneButtonEnable?()
+        }
+    }
     
+    var selectedCategory: String? {
+        didSet {
+            isDoneButtonEnable?()
+        }
+    }
     
-    func createNewTask(trackerNameTextField: String) {
-        guard let category = selectedCategory,
+    var selectedSchedule: String? {
+        didSet {
+            isDoneButtonEnable?()
+        }
+    }
+    
+    var isDoneButtonEnable: ( () -> Void )?
+        
+    func getBackToMainScreen() {
+        let cancelCreatingTrackerNotification = Notification.Name("cancelCreatingTracker")
+        NotificationCenter.default.post(name: cancelCreatingTrackerNotification, object: nil)
+    }
+    
+    func createNewTracker() {
+        guard let name = trackerName,
+              let category = selectedCategory,
               let color = selectedColor,
               let emoji = selectedEmoji,
-              let schedule = selectedSchedule else { print("Что-то пошло не так"); return }
-        
-        let name = trackerNameTextField
+              let schedule = selectedSchedule else {
+            print("Smth's going wrong here"); return
+        }
         
         let newTask = TrackerCategory(header: category,
                                       trackers: [Tracker(id: UUID(),
                                                          name: name,
                                                          color: color,
                                                          emoji: emoji,
-                                                         schedule: schedule)
-                                      ])
+                                                         schedule: schedule)])
         coreDataManager.createNewTracker(newTracker: newTask)
-        informAnotherVCofCreatingTracker?()
+        getBackToMainScreen()
     }
     
-    func isAllFieldsFilled(trackerNameTextField: Bool) -> Bool {
-        let allFieldsFilled = trackerNameTextField &&
-        selectedCategory != nil &&
-        selectedEmoji != nil &&
-        selectedColor != nil &&
-        selectedSchedule != nil
+    func isAllFieldsFilled() -> Bool {
+        let allFieldsFilled =
+            trackerName != nil &&
+            trackerName != "" &&
+            selectedCategory != nil &&
+            selectedSchedule != nil &&
+            selectedEmoji != nil &&
+            selectedColor != nil
         return allFieldsFilled
     }
-    
 }
