+++ Tracker/ViewControllers/EditingTrackerVC/Views/EditingTrackerVC+CollectionView.swift
@@ -1,33 +1,14 @@
 //
-//  CreatingNewTrackerVC+CollectionView.swift
+//  EditingTrackerVC+CollectionView.swift
 //  Tracker
 //
-//  Created by Kirill Sklyarov on 09.04.2024.
+//  Created by Kirill Sklyarov on 11.04.2024.
 //
 
 import UIKit
 
 // MARK: - UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout
-extension CreatingNewTrackerViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
-    
-    func setupEmojiCollectionView() {
-        emojiCollection.dataSource = self
-        emojiCollection.delegate = self
-        emojiCollection.register(UICollectionViewCell.self, forCellWithReuseIdentifier: "emojiCell")
-        emojiCollection.register(SupplementaryView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: "header")
-        emojiCollection.backgroundColor = .white
-        emojiCollection.isScrollEnabled = false
-    }
-    
-    func setupColorsCollectionView() {
-        colorsCollection.dataSource = self
-        colorsCollection.delegate = self
-        colorsCollection.register(UICollectionViewCell.self, forCellWithReuseIdentifier: "colorsCell")
-        colorsCollection.register(SupplementaryView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: "header")
-        colorsCollection.backgroundColor = .white
-        colorsCollection.isScrollEnabled = false
-    }
-    
+extension EditingTrackerViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
     
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         if collectionView == emojiCollection {
@@ -46,7 +27,15 @@ extension CreatingNewTrackerViewController: UICollectionViewDataSource, UICollec
             view.font = .systemFont(ofSize: 32)
             view.textAlignment = .center
             cell.addSubview(view)
+            
+            if indexPath == viewModel.emojiIndexPath {
+                cell.isSelected = true
+                cell.layer.cornerRadius = 8
+                cell.backgroundColor = UIColor(named: "textFieldBackgroundColor")
+            }
+            
             return cell
+            
         } else {
             let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "colorsCell", for: indexPath)
             let view = UIImageView(frame: CGRect(x: 0, y: 0, width: 40, height: 40))
@@ -56,30 +45,59 @@ extension CreatingNewTrackerViewController: UICollectionViewDataSource, UICollec
             cell.contentView.addSubview(view)
             view.center = CGPoint(x: cell.contentView.bounds.midX,
                                   y: cell.contentView.bounds.midY)
+            
+            if indexPath == viewModel.colorIndexPath {
+                cell.isSelected = true
+                cell.layer.borderWidth = 3
+                let cellColor = colors[indexPath.row].withAlphaComponent(0.3)
+                cell.layer.borderColor = cellColor.cgColor
+                cell.layer.cornerRadius = 8
+            }
+            
             return cell
         }
     }
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        if collectionView == emojiCollection {
-            let cell = collectionView.cellForItem(at: indexPath)
-            cell?.layer.cornerRadius = 8
-            cell?.backgroundColor = UIColor(named: "textFieldBackgroundColor")
-            viewModel.selectedEmoji = viewModel.arrayOfEmoji[indexPath.row]
-        } else {
-            let cell = collectionView.cellForItem(at: indexPath)
-            cell?.layer.borderWidth = 3
-            let colors = colorFromHexToRGB(hexColors: viewModel.arrayOfColors)
-            let cellColor = colors[indexPath.row].withAlphaComponent(0.3)
-            cell?.layer.borderColor = cellColor.cgColor
-            cell?.layer.cornerRadius = 8
-            viewModel.selectedColor = viewModel.arrayOfColors[indexPath.row]
+        guard let emojiIndexPath = viewModel.emojiIndexPath,
+              let colorIndexPath = viewModel.colorIndexPath else {
+            print("Smth's going wrong"); return
         }
-//        isCreateButtonEnable()
+        
+        if collectionView == emojiCollection {
+            if let originalEmojiCell = collectionView.cellForItem(at: emojiIndexPath) {
+                originalEmojiCell.backgroundColor = .clear
+            }
+         
+            if let cell = collectionView.cellForItem(at: indexPath) {
+                cell.layer.cornerRadius = 8
+                cell.backgroundColor = UIColor(named: "textFieldBackgroundColor")
+                viewModel.selectedEmoji = viewModel.arrayOfEmoji[indexPath.row]
+            }
+        } else {
+            if let originalColorCell = collectionView.cellForItem(at: colorIndexPath) {
+                originalColorCell.layer.borderWidth = 0
+            }
+            
+            if let cell = collectionView.cellForItem(at: indexPath) {
+                cell.layer.borderWidth = 3
+                let colors = colorFromHexToRGB(hexColors: viewModel.arrayOfColors)
+                let cellColor = colors[indexPath.row].withAlphaComponent(0.3)
+                cell.layer.borderColor = cellColor.cgColor
+                cell.layer.cornerRadius = 8
+                viewModel.selectedColor = viewModel.arrayOfColors[indexPath.row]
+            }
+        }
+        isCreateButtonEnable()
     }
     
     func collectionView(_ collectionView: UICollectionView, didDeselectItemAt indexPath: IndexPath) {
+        guard let emojiIndexPath = viewModel.emojiIndexPath else {
+            print("Smth's going wrong"); return
+        }
+        
         if collectionView == emojiCollection {
+            collectionView.cellForItem(at: emojiIndexPath)?.backgroundColor = .clear
             let cell = collectionView.cellForItem(at: indexPath)
             cell?.backgroundColor = .clear
         } else {
@@ -119,7 +137,7 @@ extension CreatingNewTrackerViewController: UICollectionViewDataSource, UICollec
         default:
             id = ""
         }
-        guard let view = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: id, for: indexPath) as? SupplementaryView else { return UICollectionReusableView ()}
+       guard let view = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: id, for: indexPath) as? SupplementaryView else { return UICollectionReusableView() }
         if collectionView == emojiCollection {
             view.label.text = "Emoji"
         } else {
@@ -132,4 +150,3 @@ extension CreatingNewTrackerViewController: UICollectionViewDataSource, UICollec
         UIEdgeInsets(top: 24, left: 0, bottom: 0, right: 0)
     }
 }
-
